Divide the conveyer to 2 subsystems
