+++ src/main/java/robot/subsystems/conveyer/Constants.java
@@ -1,4 +1,6 @@
 package robot.subsystems.conveyer;
 
 public class Constants {
+    public final static double speedFORtriserion = 0;
+    public final static double shooterSpeed = 0;
 }
