+++ src/main/java/robot/subsystems/conveyer/ConveyerShooter.java
@@ -10,40 +10,24 @@ package robot.subsystems.conveyer;
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.VictorSPX;
 import edu.wpi.first.wpilibj.command.Subsystem;
-import robot.subsystems.conveyer.Commands.TriserionCommand;
 
 /**
  * An example subsystem.  You can replace me with your own Subsystem.
  */
-public class Conveyer extends Subsystem {
+public class ConveyerShooter extends Subsystem {
     // Put methods for controlling this subsystem
     // here. Call these from Commands.
-    private final VictorSPX ConveyerRotation = new VictorSPX(Ports.ConRotation);
     private final VictorSPX ConveyerForShooter = new VictorSPX(Ports.ConToSHOOTER);
 
-    public Conveyer(){
-        ConveyerForShooter.setInverted(false);
-        ConveyerRotation.setInverted(false);
-    }
 
     @Override
     public void initDefaultCommand() {
         // Set the default command for a subsystem here.
         // setDefaultCommand(new MySpecialCommand());
-        setDefaultCommand(new TriserionCommand());
-    }
-
-    /**
-     * @auther orel
-     * @param speed the speed for the motor
-     */
-    public void SetSpeedForRotation(double speed){
-        ConveyerRotation.set(ControlMode.PercentOutput,speed);
-    }
-
-    public  void  setSpeedConveyerForShooter(double speed){
-        ConveyerForShooter.set(ControlMode.PercentOutput,speed);
 
     }
 
+    public void setSpeed(double speed) {
+        ConveyerForShooter.set(ControlMode.PercentOutput, speed);
+    }
 }
